renamed project to skateright_flutter, organized lib
