+++ skateright_flutter/lib/entities/spot.dart
