+++ skateright_flutter/lib/entities/user.dart
