+++ skateright_flutter/lib/map/fake_spot.dart
@@ -1,4 +1,4 @@
-import 'spot.dart';
+import '../entities/spot.dart';
 
 final _comments = [
   Comment(
