+++ skateright_flutter/lib/map/places_interface.dart
@@ -3,7 +3,7 @@ import 'dart:developer';
 import 'package:location/location.dart';
 import 'package:http/http.dart';
 import 'dart:convert';
-import 'spot.dart';
+import '../entities/spot.dart';
 
 class PlacesInterface {
   PlacesInterface({required this.location});
@@ -14,7 +14,7 @@ class PlacesInterface {
   /// future plans will have a variable/sliding searchRadius
   ///   - define backend calls as dynamic in searchtem & radius (with cap)
   final int searchRadius = 5000; 
-  final String apiKey = "";
+  final String apiKey = "AIzaSyBHbE8gY1lkShRnfptN5wLNJgB06qgFNvg";
   final String exactSearchPostURL =
       "https://maps.googleapis.com/maps/api/place/findplacefromtext/json";
   final String nearbyPostURL =
