+++ skateright_flutter/pubspec.yaml
@@ -1,4 +1,4 @@
-name: map_app
+name: skateright_flutter
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
