+++ skateright_flutter/web/index.html
@@ -23,13 +23,13 @@
   <!-- iOS meta tags & icons -->
   <meta name="apple-mobile-web-app-capable" content="yes">
   <meta name="apple-mobile-web-app-status-bar-style" content="black">
-  <meta name="apple-mobile-web-app-title" content="map_app">
+  <meta name="apple-mobile-web-app-title" content="skateright_flutter">
   <link rel="apple-touch-icon" href="icons/Icon-192.png">
 
   <!-- Favicon -->
   <link rel="icon" type="image/png" href="favicon.png"/>
 
-  <title>map_app</title>
+  <title>skateright_flutter</title>
   <link rel="manifest" href="manifest.json">
 </head>
 <body>
