+++ skateright_flutter/windows/runner/Runner.rc
@@ -90,12 +90,12 @@ BEGIN
         BLOCK "040904e4"
         BEGIN
             VALUE "CompanyName", "com.example" "\0"
-            VALUE "FileDescription", "map_app" "\0"
+            VALUE "FileDescription", "skateright_flutter" "\0"
             VALUE "FileVersion", VERSION_AS_STRING "\0"
-            VALUE "InternalName", "map_app" "\0"
+            VALUE "InternalName", "skateright_flutter" "\0"
             VALUE "LegalCopyright", "Copyright (C) 2022 com.example. All rights reserved." "\0"
-            VALUE "OriginalFilename", "map_app.exe" "\0"
-            VALUE "ProductName", "map_app" "\0"
+            VALUE "OriginalFilename", "skateright_flutter.exe" "\0"
+            VALUE "ProductName", "skateright_flutter" "\0"
             VALUE "ProductVersion", VERSION_AS_STRING "\0"
         END
     END
